pyproject.toml
[tool.poetry]
name = "pdf-processor"
version = "0.1.0"
description = "API to upload, process, and store PDF data using Google Cloud Document AI"
authors = ["Caio Santos"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {version = "^0.34.0", extras = ["standard"]}
python-multipart = "^0.0.18"
python-jose = {version = "^3.4.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
bcrypt = "4.0.1"
google-cloud-documentai = "^3.0.0"
google-cloud-storage = "^2.18.0"
pydantic-settings = "^2.7.0"
slowapi = "^0.1.9"
pymupdf = "^1.26.7"
google-cloud-firestore = "^2.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.25.0"
httpx = "^0.28.0"
ruff = "^0.9.0"
mypy = "^1.14.0"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
